Disable adaptive support in data_sPCA by default
If the adaptive support flag is set to true, we use the function
``choose_support_v6`` to set R and c. If it's set to false (the
default), it bypasses the function and used the full supports.
Until we can fix ``choose_support_v6``, this will enable us to get
accurate reconstructions at a certain computational and accuracy
cost.