Workaround for Octave MEX file bug
Both the CIMS NUFFT and FINUFFT use MWRAP, which has some global
variables that it uses internally. These are, for some reason,
shared between MEX files in Octave, which means that only one
function can be called at a time. To fix this, we clear the
functions before each call using `clear functions`.